+++ hodor-common/src/test/java/org/dromara/hodor/common/kv/SingleHodorKVClientTest.java
@@ -15,9 +15,8 @@
  * limitations under the License.
  */
 
-package com.dromara.hodor.common.kv;
+package org.dromara.hodor.common.kv;
 
-import java.util.Arrays;
 import org.dromara.hodor.common.raft.kv.core.HodorKVClient;
 
 /**
@@ -26,12 +25,11 @@ import org.dromara.hodor.common.raft.kv.core.HodorKVClient;
  * @author tomgs
  * @since 1.0
  */
-public class HodorKVClientTest {
+public class SingleHodorKVClientTest {
 
     public static void main(String[] args) {
-        HodorKVClient kvClient = new HodorKVClient("127.0.0.1:8081,127.0.0.1:8082,127.0.0.1:8083");
-        kvClient.defaultKvOperator()
-            .put("a".getBytes(), "1".getBytes());
+        HodorKVClient kvClient = new HodorKVClient("127.0.0.1:8081");
+        kvClient.defaultKvOperator().put("a".getBytes(), "1".getBytes());
 
         byte[] bytes = kvClient.defaultKvOperator().get("a".getBytes());
         System.out.println(new String(bytes));
@@ -39,11 +37,12 @@ public class HodorKVClientTest {
         kvClient.defaultKvOperator().delete("a".getBytes());
 
         bytes = kvClient.defaultKvOperator().get("a".getBytes());
-        System.out.println(Arrays.toString(bytes));
+        System.out.println(bytes);
 
         final Boolean aBoolean = kvClient.defaultKvOperator().containsKey("a".getBytes());
         System.out.println(aBoolean);
 
+        kvClient.defaultKvOperator().put("b".getBytes(), "2".getBytes());
         final Boolean bBoolean = kvClient.defaultKvOperator().containsKey("b".getBytes());
         System.out.println(bBoolean);
     }
