+++ hodor-common/src/test/java/org/dromara/hodor/common/raft/Demo2HodorRaftStateMachine.java
@@ -15,12 +15,11 @@
  * limitations under the License.
  */
 
-package com.dromara.hodor.common.raft;
+package org.dromara.hodor.common.raft;
 
 import java.io.IOException;
 import java.util.Collection;
 import java.util.concurrent.CompletableFuture;
-import org.apache.ratis.proto.RaftProtos;
 import org.apache.ratis.protocol.Message;
 import org.apache.ratis.protocol.RaftGroupId;
 import org.apache.ratis.protocol.RaftGroupMemberId;
@@ -36,14 +35,11 @@ import org.dromara.hodor.common.raft.HodorRaftStateMachine;
  * @author tomgs
  * @since 1.0
  */
-public class DemoHodorRaftStateMachine extends HodorRaftStateMachine {
-
-    private RaftServer raftServer;
+public class Demo2HodorRaftStateMachine extends HodorRaftStateMachine {
 
     @Override
     public void initialize(RaftServer raftServer, RaftGroupId raftGroupId, RaftStorage storage) throws IOException {
         super.initialize(raftServer, raftGroupId, storage);
-        this.raftServer = raftServer;
     }
 
     @Override
@@ -63,16 +59,8 @@ public class DemoHodorRaftStateMachine extends HodorRaftStateMachine {
 
     @Override
     public void notifyLeaderChanged(RaftGroupMemberId groupMemberId, RaftPeerId newLeaderId) {
-        System.out.println("DemoHodorRaftStateMachine#notifyLeaderChanged ================================ " + groupMemberId + ":" + newLeaderId);
+        System.out.println("Demo2HodorRaftStateMachine#notifyLeaderChanged ================================ " + groupMemberId + ":" + newLeaderId);
         RaftPeerId currentPeerId = groupMemberId.getPeerId();
-        try {
-            final RaftServer.Division division = raftServer.getDivision(getGroupId());
-            final boolean leader = division.getInfo().isLeader();
-            System.out.println("IS_LEADER:" + leader);
-            final RaftProtos.RoleInfoProto roleInfoProto = division.getInfo().getRoleInfoProto();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
         if (currentPeerId.equals(newLeaderId)) {
             System.out.println("LEADER");
         } else {
@@ -83,8 +71,6 @@ public class DemoHodorRaftStateMachine extends HodorRaftStateMachine {
     @Override
     public void notifyNotLeader(Collection<TransactionContext> pendingEntries) throws IOException {
         // 不是主节点时回调
-        System.out.println("DemoHodorRaftStateMachine#notifyNotLeader ================================ " + pendingEntries);
+        System.out.println("Demo2HodorRaftStateMachine#notifyNotLeader ================================ " + pendingEntries);
     }
-
-
 }
