+++ hodor-extension/hodor-extension-cache-redis-impl/src/main/java/org/dromara/hodor/common/storage/cache/impl/RedisCacheSource.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.dromara.hodor.common.storage.cache.impl;
+package org.dromara.hodor.common.storage.cache.impl;
 
 import org.dromara.hodor.common.extension.Join;
 import org.dromara.hodor.common.storage.cache.CacheClient;
