+++ hodor-extension/hodor-extension-cache-redis-impl/src/main/resources/META-INF/hodor/org.dromara.hodor.common.storage.cache.HodorCacheSource
@@ -1,1 +1,1 @@
-redis=com.dromara.hodor.common.storage.cache.impl.RedisCacheSource
+redis=org.dromara.hodor.common.RedisCacheSource
